HBServer -> HBHTTPServer
Rename protocol HBServer to HBHTTPServer
HBHTTPServer is not HBPosixHTTPServer
Moved more code into the protocol
